fix 1

+++ src/main/java/com/ao/auditorium/web/controller/pages/my/lecturing/MyLecturingCoursesPageController.java
@@ -8,12 +8,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
 
 import javax.annotation.Resource;
-import java.util.Optional;
 
 @Controller()
 public class MyLecturingCoursesPageController {
@@ -21,24 +20,20 @@ public class MyLecturingCoursesPageController {
     private CourseRepository courseRepository;
 
     @GetMapping("/my/lecturing-courses")
-    public String showAllCourses(Model model,@RequestParam("status") Optional<String> status) {
+    public String showAllCourses(Model model) {
         model.addAttribute("courses", courseRepository.findAll());
-        if (status.isPresent())
-            model.addAttribute("status", status.get());
-        else
-            model.addAttribute("status", null);
         return WebConstants.Pages.MY_LECTURING_FOLDER+"lecturingCourses";
     }
 
     @GetMapping("/my/lecturing-courses/")
-    public String createCourses(Model model) {
+    public String createCourse(Model model) {
         model.addAttribute("course", null);
         model.addAttribute("courseId", 0 );
         return WebConstants.Pages.MY_LECTURING_FOLDER+"lecturingCourse";
     }
 
     @GetMapping("/my/lecturing-courses/{courseId}")
-    public String updateCourses(@PathVariable String courseId, Model model) {
+    public String updateCourse(@PathVariable String courseId, Model model) {
         Long num;
         try
         {
@@ -56,25 +51,26 @@ public class MyLecturingCoursesPageController {
     }
 
     @PostMapping("/my/lecturing-courses/{courseId}")
-    public RedirectView redirect(@PathVariable Long courseId,RedirectAttributes attributes,@RequestParam("courseName") String code,@RequestParam("courseName") String name,@RequestParam("courseDescription") String description){
+    public RedirectView saveCourse(@PathVariable Long courseId,RedirectAttributes attributes,@ModelAttribute CourseForm courseForm){
         if (courseRepository.findById(courseId).isPresent()) {
             Course course = courseRepository.findById(courseId).get();
-            course.setName(name);
-            course.setCode(code);
-            course.setDescription(description);
+            course.setName(courseForm.getName());
+            course.setCode(courseForm.getCode());
+            course.setDescription(courseForm.getDescription());
             courseRepository.save(course);
         }else{
-            Course newCourse = new Course(code,name,description);
+            Course newCourse = new Course(courseForm.getName(),courseForm.getCode(),courseForm.getDescription());
             courseRepository.save(newCourse);
         }
-        attributes.addAttribute("status", "success");
+        attributes.addFlashAttribute("status", "success");
         return new RedirectView("/my/lecturing-courses");
     }
 
     @PostMapping("/my/lecturing-courses/{courseId}/delete")
-    public RedirectView redirect(@PathVariable Long courseId,RedirectAttributes attributes){
-        courseRepository.deleteById(courseId);
-        attributes.addAttribute("status", "success");
+    public RedirectView deleteCourse(@PathVariable Long courseId,RedirectAttributes attributes){
+        if (courseRepository.findById(courseId).isPresent())
+            courseRepository.deleteById(courseId);
+        attributes.addFlashAttribute("status", "success");
         return new RedirectView("/my/lecturing-courses");
     }
